+++ packages/react-google-maps-api-infobox/src/InfoBox.tsx
@@ -13,6 +13,7 @@ export class InfoBox {
   boxStyle: {
     [key: string]: any
   }
+
   closeBoxMargin: string
   closeBoxURL: string
   infoBoxClearance: google.maps.Size
@@ -105,6 +106,7 @@ export class InfoBox {
         this.div.appendChild(this.content)
       }
 
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       const panes = this.getPanes()
       panes[this.pane].appendChild(this.div) // Add the InfoBox div to the DOM
@@ -233,6 +235,7 @@ export class InfoBox {
 
   panBox(disablePan?: boolean): void {
     if (this.div && !disablePan) {
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       const map: google.maps.Map | google.maps.StreetViewPanorama | null | undefined = this.getMap()
 
@@ -249,8 +252,10 @@ export class InfoBox {
         }
 
         const mapDiv = map.getDiv()
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
         const mapWidth = mapDiv.offsetWidth
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
         const mapHeight = mapDiv.offsetHeight
         const iwOffsetX = this.pixelOffset.width
@@ -260,6 +265,7 @@ export class InfoBox {
         const padX = this.infoBoxClearance.width
         const padY = this.infoBoxClearance.height
 
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
         const projection: google.maps.MapCanvasProjection = this.getProjection()
         const pixPosition = projection.fromLatLngToContainerPixel(this.position)
@@ -305,7 +311,9 @@ export class InfoBox {
       // Apply style values defined in the boxStyle parameter:
       const boxStyle = this.boxStyle
       for (const i in boxStyle) {
-        if (boxStyle.hasOwnProperty(i)) {
+
+        if (Object.prototype.hasOwnProperty.call(boxStyle, i)) {
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
           // @ts-ignore
           this.div.style[i] = boxStyle[i]
         }
@@ -320,6 +328,7 @@ export class InfoBox {
         // See http://www.quirksmode.org/css/opacity.html
         const opacity = parseFloat(this.div.style.opacity || '')
 
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
         this.div.style.msFilter =
           '"progid:DXImageTransform.Microsoft.Alpha(Opacity=' + opacity * 100 + ')"'
@@ -360,9 +369,11 @@ export class InfoBox {
         bw.right = parseInt(computedStyle.borderRightWidth || '', 10) || 0
       }
     } else if (
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       document.documentElement.currentStyle // MSIE
     ) {
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       const currentStyle = this.div.currentStyle
 
@@ -389,8 +400,10 @@ export class InfoBox {
     this.createInfoBoxDiv()
 
     if (this.div) {
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       const projection: google.maps.MapCanvasProjection = this.getProjection()
+
       const pixPosition = projection.fromLatLngToDivPixel(this.position)
 
       if (pixPosition !== null) {
@@ -561,8 +574,10 @@ export class InfoBox {
   }
 
   getVisible(): boolean {
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
     const map: google.maps.Map | google.maps.StreetViewPanorama | null | undefined = this.getMap()
+
     let isVisible
 
     if (typeof map === 'undefined' || map === null) {
@@ -593,14 +608,18 @@ export class InfoBox {
     anchor?: google.maps.MVCObject
   ): void {
     if (anchor) {
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       this.position = anchor.getPosition()
+
       this.moveListener = google.maps.event.addListener(
         anchor,
         'position_changed',
         () => {
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
           // @ts-ignore
           const position = anchor.getPosition()
+
           this.setPosition(position)
         }
       )
@@ -609,12 +628,14 @@ export class InfoBox {
         anchor,
         'map_changed',
         () => {
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
           // @ts-ignore
           this.setMap(anchor.map)
         }
       )
     }
 
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
     this.setMap(map)
 
@@ -651,6 +672,7 @@ export class InfoBox {
       this.contextListener = null
     }
 
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
     this.setMap(null)
   }
@@ -659,13 +681,16 @@ export class InfoBox {
     return function applyExtend(object: any) {
       // eslint-disable-next-line guard-for-in
       for (const property in object.prototype) {
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
-        if (!this.prototype.hasOwnProperty(property)) {
+        if (!Object.prototype.hasOwnProperty.call(this, property)) {
+          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
           // @ts-ignore
           this.prototype[property] = object.prototype[property]
         }
       }
 
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       return this
     }.apply(obj1, [obj2])