+++ packages/react-google-maps-api/src/components/drawing/Data.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react'
+import { PureComponent } from 'react'
 
 import { unregisterEvents, applyUpdatersToPropsAndRegisterEvents } from '../../utils/helper'
 
@@ -119,7 +119,7 @@ export interface DataProps {
   onUnmount?: ((data: google.maps.Data) => void) | undefined
 }
 
-export class Data extends React.PureComponent<DataProps, DataState> {
+export class Data extends PureComponent<DataProps, DataState> {
   static contextType = MapContext
 
   registeredEvents: google.maps.MapsEventListener[] = []