+++ packages/react-google-maps-api/src/index.ts
@@ -1,83 +1,83 @@
-export { default as GoogleMap, GoogleMapProps } from './GoogleMap'
+export { default as GoogleMap, type GoogleMapProps } from './GoogleMap'
 
-export { default as LoadScript, LoadScriptProps } from './LoadScript'
+export { default as LoadScript, type LoadScriptProps } from './LoadScript'
 
-export { default as LoadScriptNext, LoadScriptNextProps } from './LoadScriptNext'
+export { default as LoadScriptNext, type LoadScriptNextProps } from './LoadScriptNext'
 
 export { useLoadScript } from './useLoadScript'
 
 export { useJsApiLoader } from './useJsApiLoader'
 
-export { default as TrafficLayer, TrafficLayerProps } from './components/maps/TrafficLayer'
+export { default as TrafficLayer, TrafficLayerF, type TrafficLayerProps } from './components/maps/TrafficLayer'
 
-export { default as BicyclingLayer, BicyclingLayerProps } from './components/maps/BicyclingLayer'
+export { default as BicyclingLayer, BicyclingLayerF, type BicyclingLayerProps } from './components/maps/BicyclingLayer'
 
-export { default as TransitLayer, TransitLayerProps } from './components/maps/TransitLayer'
+export { default as TransitLayer, TransitLayerF, type TransitLayerProps } from './components/maps/TransitLayer'
 
-export { default as DrawingManager, DrawingManagerProps } from './components/drawing/DrawingManager'
+export { default as DrawingManager, DrawingManagerF, type DrawingManagerProps } from './components/drawing/DrawingManager'
 
-export { default as Marker, MarkerProps } from './components/drawing/Marker'
+export { default as Marker, type MarkerProps } from './components/drawing/Marker'
 
 export {
   default as MarkerClusterer,
-  ClustererProps as MarkerClustererProps,
+  type ClustererProps as MarkerClustererProps,
 } from './components/addons/MarkerClusterer'
 
-export { default as InfoBox, InfoBoxProps } from './components/addons/InfoBox'
+export { default as InfoBox, InfoBoxF, type InfoBoxProps } from './components/addons/InfoBox'
 
-export { default as GoogleMarkerClusterer, GoogleMarkerClustererProps } from './components/addons/GoogleMarkerClusterer'
+export { default as GoogleMarkerClusterer, type GoogleMarkerClustererProps } from './components/addons/GoogleMarkerClusterer'
 
-export { default as InfoWindow, InfoWindowProps } from './components/drawing/InfoWindow'
+export { default as InfoWindow, InfoWindowF, type InfoWindowProps } from './components/drawing/InfoWindow'
 
-export { default as Polyline, PolylineProps } from './components/drawing/Polyline'
+export { default as Polyline, type PolylineProps } from './components/drawing/Polyline'
 
-export { default as Polygon, PolygonProps } from './components/drawing/Polygon'
+export { default as Polygon, type PolygonProps } from './components/drawing/Polygon'
 
-export { default as Rectangle, RectangleProps } from './components/drawing/Rectangle'
+export { default as Rectangle, type RectangleProps } from './components/drawing/Rectangle'
 
-export { default as Circle, CircleProps } from './components/drawing/Circle'
+export { default as Circle, type CircleProps } from './components/drawing/Circle'
 
-export { default as Data, DataProps } from './components/drawing/Data'
+export { default as Data, type DataProps } from './components/drawing/Data'
 
-export { default as KmlLayer, KmlLayerProps } from './components/kml/KmlLayer'
+export { default as KmlLayer, type KmlLayerProps } from './components/kml/KmlLayer'
 
-export { default as OverlayView, OverlayViewProps } from './components/dom/OverlayView'
+export { default as OverlayView, type OverlayViewProps } from './components/dom/OverlayView'
 
-export { default as GroundOverlay, GroundOverlayProps } from './components/overlays/GroundOverlay'
+export { default as GroundOverlay, type GroundOverlayProps } from './components/overlays/GroundOverlay'
 
-export { default as HeatmapLayer, HeatmapLayerProps } from './components/heatmap/HeatmapLayer'
+export { default as HeatmapLayer, type HeatmapLayerProps } from './components/heatmap/HeatmapLayer'
 
 export {
   default as StreetViewPanorama,
-  StreetViewPanoramaProps,
+  type StreetViewPanoramaProps,
 } from './components/streetview/StreetViewPanorama'
 
 export {
   default as StreetViewService,
-  StreetViewServiceProps,
+  type StreetViewServiceProps,
 } from './components/streetview/StreetViewService'
 
 export {
   default as DirectionsService,
-  DirectionsServiceProps,
+  type DirectionsServiceProps,
 } from './components/directions/DirectionsService'
 
 export {
   default as DirectionsRenderer,
-  DirectionsRendererProps,
+  type DirectionsRendererProps,
 } from './components/directions/DirectionsRenderer'
 
 export {
   default as DistanceMatrixService,
-  DistanceMatrixServiceProps,
+  type DistanceMatrixServiceProps,
 } from './components/distance-matrix/DistanceMatrixService'
 
 export {
   default as StandaloneSearchBox,
-  StandaloneSearchBoxProps,
+  type StandaloneSearchBoxProps,
 } from './components/places/StandaloneSearchBox'
 
-export { default as Autocomplete, AutocompleteProps } from './components/places/Autocomplete'
+export { default as Autocomplete, type AutocompleteProps } from './components/places/Autocomplete'
 
 export { default as MapContext, useGoogleMap } from './map-context'