+++ packages/react-google-maps-api/src/map-context.ts
@@ -6,7 +6,7 @@ const MapContext = createContext<google.maps.Map | null>(null)
 export function useGoogleMap(): google.maps.Map | null {
   invariant(!!useContext, 'useGoogleMap is React hook and requires React version 16.8+')
 
-  const map = useContext(MapContext)
+  const map = useContext<google.maps.Map | null>(MapContext)
 
   invariant(!!map, 'useGoogleMap needs a GoogleMap available up in the tree')